requirements.txt
Django==1.6.1
Fabric==1.8.0
Fiona==1.0.2
Jinja2==2.7.1
MarkupSafe==0.18
Mezzanine==3.0.5
Pygments==1.6
Shapely==1.2.18
South==0.8.4
amqp==1.4.0
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
arrow==0.4.2
bagit==1.3.6
billiard==3.3.0.13
bleach==1.4
celery==3.1.13
certifi==0.0.8
chardet==2.0.1
django-appconf==0.6
django-celery==3.1.1
django-compressor==1.3
django-crispy-forms==1.4.0
django-debug-toolbar==0.11.0
django-extensions==1.2.5
django-nose==1.2
django-picklefield==0.3.1
django-tastypie==0.10.0
django-tastypie-swagger==0.1.2
django-timedeltafield==0.7.0
docutils==0.11
ecdsa==0.10
filebrowser-safe==0.3.1
future==0.9.0
geojson==1.0.5
grappelli-safe==0.3.4
html5lib==1.0b3
imposm==2.5.0
imposm.parser==1.0.5
ipython==1.1.0
kombu==3.0.21
lxml==3.2.4
matplotlib==1.3.1
mimeparse==0.1.3
nose==1.3.0
numexpr==2.2.2
oauthlib==0.6.0
paramiko==1.12.0
protobuf==2.5.0
psycopg2==2.5.1
pyasn1==0.1.7
pycrypto==2.6.1
pyparsing==2.0.1
pyproj==1.9.3
pysqlite==2.6.3
python-dateutil==2.2
python-safe==1.1.2
pytz==2013.9
pyzmq==14.1.1
readline==6.2.4.1
redis==2.10.3
requests
requests-oauthlib
rsa==3.1.2
sh==1.09
simplejson==3.3.1
six==1.5.2
sqlparse==0.1.10
ssh==1.8.0
ssh-import-id==3.14
tables==3.0.0
tornado==3.1.1
uWSGI==1.9.20
urllib3==1.5
wsgiref==0.1.2
google.foresite-toolkit
pandas
suds-jurko
